pyproject.toml
[tool.poetry]
name = "torah-stream-api"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "torah_stream_api"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.3"
strawberry-graphql = "^0.142.2"
strawberry-graphql-django = "^0.8.2"
strawberry-django-plus = "^1.31.1"
django-easy-audit = "^1.3.3"
psycopg2-binary = "^2.9.5"
django-tree-queries = "^0.12.0"
djangoql = "^0.17.1"


[tool.poetry.group.dev.dependencies]
django-types = "^0.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
